YOLO-v3-using-PyTorch-implementation/requirements.txt
numpy
torchvision
matplotlib
tensorflow
tensorboard
terminaltables
pillow
tqdm
